pyproject.toml
[project]
name = "chainlit-and-outline"
version = "0.0.1"
description = "This is a template repository for Python projects that use uv for their dependency management."
authors = [{ name = "Mostafa Kazemi", email = "mkay1375@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.9,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "chainlit>=2.3.0",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic-ai>=0.1.11",
]

[project.urls]
Homepage = "https://mkay1375.github.io/chainlit-and-outline/"
Repository = "https://github.com/mkay1375/chainlit-and-outline"
Documentation = "https://mkay1375.github.io/chainlit-and-outline/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",

    "mypy>=0.991",

    "ruff>=0.11.5",

]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"



[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # Loop control variable `message_history` not used within loop body
    "B007",
    #  Loop control variable `message_history` overrides iterable it iterates
    "B020",
    # Avoid specifying long messages outside the exception class
    "TRY003",
    # Consider moving this statement to an `else` block
    "TRY300",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true
